Inplace updates + smaller formulations
ngraph/ - Added EmbeddedLookupBackprop so that embedding table backprop
is actually somewhat sane now.

nGraph.jl - Initial support for "in place" updating of embedding tables.

AutoTM.jl - Detect if tensors are fixed in a single memory pool. If so,
we now avoid generating a bunch of useless variables in the ILP.
This greatly decreases ILP model size if, for example, nGraph function
parameters are fixed in either DRAM or PMM.

It should be noted that originally, argument tensors were special cased,
for this behavior - but now it applies to arbitrary tensors.